Format tpm2, and move constants to dedicated submodule
